+++ src/db/db.ts
@@ -1,4 +1,5 @@
 import mysql, { Pool } from 'mysql2/promise';
+import { minuteMilliseconds } from '../util/constants';
 
 export const dbPool: Pool = mysql.createPool({
   host: process.env.DATABASE_HOST,
@@ -7,8 +8,8 @@ export const dbPool: Pool = mysql.createPool({
   database: process.env.DATABASE_NAME,
   waitForConnections: true,
   connectionLimit: 50,
-  idleTimeout: 1000 * 60 * 5,
-  queueLimit: 50,
+  idleTimeout: minuteMilliseconds * 5,
+  queueLimit: 100,
   enableKeepAlive: true,
   keepAliveInitialDelay: 0,
   multipleStatements: true,