+++ src/util/globalUtils.ts
@@ -10,3 +10,16 @@ function getMonthName(date: Date): string {
 function getTime(date: Date): string {
   return new Intl.DateTimeFormat('en-GB', { timeStyle: 'short' }).format(date);
 };
+
+export function containsInvalidWhitespace(string: string): boolean {
+  if (string.trim() !== string) {
+    return true;
+  };
+
+  const doubleWhitespacesRemoved: string = string.split(' ').filter((char: string) => char !== '').join(' ');
+  if (string !== doubleWhitespacesRemoved) {
+    return true;
+  };
+
+  return false;
+};